move the outer edge merge handler to a different package
