+++ resotocore/resotocore/action_handlers/merge_outer_edge_handler.py
